+++ src/lib.rs
@@ -1,15 +1,18 @@
 pub mod matrix;
+pub mod vector;
+pub mod stat;
 
 #[macro_use]
-pub mod vector_macro;
-
-pub mod stat;
+pub mod r_macro;
 
 #[allow(unused_imports)]
 pub use matrix::*;
 
 #[allow(unused_imports)]
-pub use vector_macro::*;
+pub use vector::*;
 
 #[allow(unused_imports)]
 pub use stat::*;
+
+#[allow(unused_imports)]
+pub use r_macro::*;