+++ src/vector.rs
@@ -3,96 +3,6 @@ use std::f64::{MIN};
 
 pub type Vector = Vec<f64>;
 
-/// R like concatenate (Type: Vec\<f64\>)
-///
-/// # Examples
-/// ```
-/// extern crate peroxide;
-/// use peroxide::*;
-///
-/// let a = c![1,2,3,4];
-/// let b = c![5,6,7,8];
-/// let c = c![a; b];
-/// println!("{:?}", a); // [1,2,3,4]
-/// println!("{:?}", b); // [5,6,7,8]
-/// println!("{:?}", c); // [1,2,3,4,5,6,7,8]
-/// ```
-#[macro_export]
-macro_rules! c {
-    ( $( $x:expr ),* ) => {
-        {
-            let mut v: Vec<f64> = Vec::new();
-            let mut l: usize = 0;
-            $(
-                v.push($x as f64);
-                l += 1;
-            )*
-            v
-        }
-    };
-    ( $( $x:expr );* ) => {
-        {
-            let mut v: Vec<f64> = Vec::new();
-            $(
-                v.extend(&$x);
-            )*
-            v
-        }
-    }
-}
-
-/// R like seq macro
-///
-/// # Examples
-/// ```
-/// extern crate peroxide;
-/// use peroxide::*;
-///
-/// assert_eq!(seq!(1,10,1), c!(1,2,3,4,5,6,7,8,9,10));
-/// assert_eq!(seq!(1,10,1), seq!(1;10;1));
-/// ```
-#[macro_export]
-macro_rules! seq {
-    ( $start:expr, $end:expr, $step:expr ) => {
-        {
-            let s = $start as f64;
-            let e = $end as f64;
-            let step = $step as f64;
-
-            assert!(e > s);
-
-            let factor: f64 = (e - s) / step;
-            let l: usize = factor as usize + 1;
-            let mut v: Vec<f64> = Vec::new();
-
-            for i in 0 .. l {
-                v.push(s + step * (i as f64));
-            }
-            v
-        }
-    };
-    ( $start:expr; $end:expr; $step:expr ) => {
-        seq!($start, $end, $step)
-    }
-}
-
-/// zeros - like numpy
-///
-/// # Examples
-/// ```
-/// extern crate peroxide;
-/// use peroxide::*;
-///
-/// let a = zeros!(4);
-/// assert_eq!(a, c!(0,0,0,0));
-/// ```
-#[macro_export]
-macro_rules! zeros {
-    ( $n:expr ) => {
-        vec![0f64; $n]
-    };
-}
-
 pub trait FPVector {
     fn fmap<F>(&self, f: F) -> Vector where F: Fn(f64) -> f64;
     fn reduce<F, T>(&self, init: T, f: F) -> f64
@@ -154,7 +64,7 @@ impl Algorithm for Vector {
     /// ```
     fn rank(&self) -> Vec<usize> {
         let l = self.len();
-        let idx = seq!(1;l;1).into_iter().map(|x| x as usize).collect::<Vec<usize>>();
+        let idx = (1 .. (l+1)).map(|x| x as usize).collect::<Vec<usize>>();
 
         let mut vec_tup = self.clone().into_iter().zip(idx.clone()).collect::<Vec<(f64, usize)>>();
         vec_tup.sort_by(|x,y| x.0.partial_cmp(&y.0).unwrap().reverse());